Add ZPIOS example
Until now there has been really no documentation for the ZPIOS
test case.  The example-zpios page is an attempt to minimally
describe the kernel module, utility, and wrapper scripts.  Along
with the component descriptions a trivial example is provided.

This commit also updates to index to reference the new ZPIOS
example page.