Remove vmtruncate() call from the zfs_setattr()
Linux kenrel uses truncate_setsize() to setattr() to truncate the file
size to a given size. This change is simplifies the code. See Linux
kernel commit 2c27c65ed0 for details. Because we care about the compatibility
with the older kernels, we can't simply replace the vmtruncate() call with
truncate_setsize(). Therefore, a function zfs_truncate_setsize() is added
which is very similar to truncate_setsize().

Signed-off-by: Prasad Joshi <prasadjoshi124@gmail.com>
Signed-off-by: Prasad Joshi <prasad@kqinfotech.com>