zfs: Tear down the mmap region and ensure it is flushed during inode eviction
The change fixes the Bug #255.

The inode eviction should unmap the pages associated with the inode. These
pages should also be flushed to disk to avoid the data loss. Therefore, use
truncate_setsize() in evict_inode() to release the pagecache.

The API truncate_setsize() was added in 2.6.35 kernel. To ensure
compatibility with the old kernel, the patch defines it's own
truncate_setsize function.

Signed-off-by: Prasad Joshi <pjoshi@stec-inc.com>