Override dh_shlibdeps warnings.
The default dh_shlibdeps rule emits hundreds of these warnings:

* symbol $sym used by $binary found in none of the libraries.

* dependency on $library could be avoided if $binaries were not
  uselessly linked against it (they use none of its symbols).

Disable these warnings through a dpkg-dhlibddeps parameter
because they are scary, spurious, and too noisy to be helpful.

Satisfying these warnings is difficult because a cmd/ utility can
depend on a library, and both can depend on another library, which
seems to be a corner case for the dependency generator.