Implement the truncate_range() inode operation.
This operation allows "hole punching" in ZFS files. On Solaris this is done
via the vop_space() system call, which maps to the zfs_space() function. So
we just need to write zpl_truncate_range() as a wrapper around zfs_space().

Note that this only works for regular files, not ZVOLs.

This is currently an insecure implementation without permission checking,
although this isn't that big of a deal since truncate_range() isn't even
callable from userspace.

This is related to issue #334.