Support the fallocate() file operation.
Currently only the (FALLOC_FL_PUNCH_HOLE) flag combination is supported,
since it's the only one that matches the behavior of zfs_space(). This
makes it pretty much useless in its current form, but it's a start.

To support other flag combinations we would need to modify zfs_space()
to make it more flexible, or emulate the desired functionality in
zpl_fallocate().

This is related to issue #334.