Initialize tqent_flags to zero in taskq_dispatch
The tqent_flags field of an allocated taskq_ent structure was not
explicitly getting initialized to zero in the taskq_dispatch function.
As a result, if the garbage that filled that field in a newly allocated
structure happened to have the bit set corresponding to the
TQENT_FLAG_PREALLOC flag, this would cause problems.

In the above scenario, the structure would get allocated in
taskq_dispatch and tq->tq_nalloc would be incremented just fine. The
problem arises in taskq_thread, as it would not be freed since the
TQENT_FLAG_PREALLOC is set (even though the structure was never
preallocated).

In the best case, this condition would get caught by an ASSERT in
taskq_destory as tq->freelist would deplete to a length of zero,
but tq->tq_nalloc would still be non zero. In the worst case, the ASSERT
would get compiled out and the function would spin, continuously
allocating and then immediately freeing taskq_ent structures.