Fix zfs_rename to update symlinks under /dev
When a zvol consisting of snapshots is renamed, the symlinks under
/dev for snapshots are not renamed at all. Exporting and reimporting
the zpool updates the link properly but that is not desirable to
do. We need to update udev in zfs_rename to update these symlinks as
part of zfs rename.

I've update the zfs_rename function to remove symlinks prior to
renaming the volume and create new links upon success or restore old
links upon failure. Also, I've moved parentname and zhrp variable
asignments up towards the beginning as they are needed for both
recursive and volume rename cases. I've also added supporting
functions zfs_create_link_snapvol and zfs_remove_link_snapvol to be
used as callbacks for zfs_iter_snapshots