Add overlay(-O) mount option support
Linux supports mounting over non-empty directories by default. In
Solaris this is not the case and -O option is required for zfs mount
to mount a zfs filesystem over a non-empty directory. For
compatibility, I've added support for -O option to mount zfs
filesystems over non-empty directories if the user wants to, just like
in Solaris.

I've defined MS_OVERLAY to record it in the flags variable if -O
option is supplied. the flags variable passes through a few functions
and its checked before performing the empty directory check in
zfs_mount function. if -O is given, the check is not performed.